sidecar-testing/flask/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
-e ../../sdks/python
